Add support for nested routes via react-router-config
This change adds support for nested routes via the react-router-config
library. This feature can be used to implement layout components to wrap
routes.

At the moment, only for the most inner route that is matched,
`loadInitialProps()` is called as we need to work out how this mechanism
will work for multiple components.